requirements.txt
amqp
arabic-reshaper
asgiref
asn1crypto
attrs
billiard
Brotli
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
coloredlogs
cron-descriptor
cryptography
cssselect2
defusedxml
Django
django-auditlog
django-celery-beat
django-cors-headers
django-debug-toolbar
django-extensions
django-filter
django-simple-history
django-templated-mail
django-timezone-field
djangorestframework
djangorestframework-simplejwt
djoser
drf-spectacular
et-xmlfile
factory-boy
Faker
flake8
fonttools
gevent
greenlet
html5lib
humanfriendly
idna
inflection
jsonschema
jsonschema-specifications
kombu
lxml
mccabe
numpy
oauthlib
opencv-python
openpyxl
oscrypto
pandas
Pillow
prompt-toolkit
psycopg2-binary
pycodestyle
pycparser
pydyf
pyflakes
pyHanko
pyhanko-certvalidator
PyJWT
pypdf
pyphen
pypng
pyreadline3
pyrsistent
python-barcode
python-bidi
python-crontab
python-dateutil
python-decouple
python-dotenv
python3-openid
pytz
PyYAML
pyzbar
qrcode
referencing
reportlab
requests
requests-oauthlib
rpds-py
six
social-auth-app-django
social-auth-core
SQLAlchemy
sqlparse
svglib
tablib
tenacity
tinycss2
typing_extensions
tzdata
tzlocal
uritemplate
uritools
urllib3
vine
wcwidth
weasyprint
webencodings
xhtml2pdf
zope.event
zope.interface
zopfli
black
python-json-logger
bleach